feat: json para criar movimentacao

+++ investment-manager/pkg/api/v1/request/movimentacao_request.go
@@ -4,6 +4,6 @@ type MovimentacaoRequest struct {
 	Data          string
 	Codigo        string
 	TipoAtivo     string
-	Quantidade    int
+	Quantidade    float64
 	ValorUnitario float64
 }
