+++ investment-manager/pkg/api/v1/response/movimentacao_response.go
@@ -10,12 +10,12 @@ type MovimentacaoResponse struct {
 	Operacao      string
 	Codigo        string
 	TipoAtivo     string
-	Quantidade    int
+	Quantidade    float64
 	ValorUnitario float64
 	ValorTotal    float64
 }
 
-func NewMovimentacaoResponse(id uint, data time.Time, operacao string, codigo string, tipoAtivo string, quantidade int,
+func NewMovimentacaoResponse(id uint, data time.Time, operacao string, codigo string, tipoAtivo string, quantidade float64,
 	valorUnitario float64, valorTotal float64) MovimentacaoResponse {
 	return MovimentacaoResponse{
 		ID:            id,
